[wip] add kmeans vertical distribution

+++ src/transforms/traits.rs
@@ -1,17 +1,17 @@
 pub trait BuildVQS {
-    fn nvrt(&self) -> u8;
-    fn bottom_level_indices(&self) -> Vec<u8>;
+    fn nvrt(&self) -> usize;
+    fn bottom_level_indices(&self) -> Vec<usize>;
     fn iter_level_values(&self) -> IterLevelValues;
-    fn values_at_level(&self, level: u8) -> Vec<f64>;
+    fn values_at_level(&self, level: usize) -> Vec<f64>;
 }
 
 pub struct IterLevelValues<'a> {
     vqs: &'a dyn BuildVQS,
-    level: u8,
+    level: usize,
 }
 
 impl<'a> Iterator for IterLevelValues<'a> {
-    type Item = (u8, Vec<f64>);
+    type Item = (usize, Vec<f64>);
 
     fn next(&mut self) -> Option<Self::Item> {
         let values = self.vqs.values_at_level(self.level);
